remove fun from extra libraries and generator callback support
