+++ src/bindings/callback.rs
@@ -5,13 +5,7 @@ use mlua::prelude::*;
 use lazy_static::lazy_static;
 use std::sync::RwLock;
 
-use self::{function::LuaFunctionCallback, generator::LuaGeneratorCallback};
 use crate::{time::BeatTimeBase, PulseIterItem};
-
-// -------------------------------------------------------------------------------------------------
-
-mod function;
-mod generator;
 
 // -------------------------------------------------------------------------------------------------
 
@@ -57,61 +51,98 @@ pub fn clear_lua_callback_errors() {
 
 // -------------------------------------------------------------------------------------------------
 
-/// Construct [`LuaCallback`] instances from a Lua function or fun generator.
-pub(crate) struct LuaCallbackFactory {}
-
-impl LuaCallbackFactory {
-    /// Checks whether the table looks like a fun generator.
-    pub fn is_fun_generator(table: &LuaTable) -> bool {
-        table
-            .contains_key("gen")
-            .and(table.contains_key("param"))
-            .and(table.contains_key("state"))
-            .unwrap_or(false)
-    }
-
-    /// Tries creating a new Lua callback from a Lua function.
-    pub fn from_function(lua: &Lua, function: LuaFunction) -> LuaResult<Box<dyn LuaCallback>> {
-        Ok(Box::new(LuaFunctionCallback::new(lua, function)?))
-    }
-
-    /// Tries creating a new Lua callback from a fun generator table.
-    pub fn from_fun_generator(lua: &Lua, table: LuaTable) -> LuaResult<Box<dyn LuaCallback>> {
-        Ok(Box::new(LuaGeneratorCallback::new(lua, table)?))
-    }
+/// Lazily evaluates a lua function the first time it's called, to either use it as a iterator,
+/// a function which returns a function, or directly as it is.
+///
+/// When calling the function the signature of the function is `fn(context): LuaResult`;
+/// The passed context is created as an empty table with the callback, and should be filled up
+/// with values before it's called.
+///
+/// Errors from callbacks should be handled by calling `self.handle_error` so external clients
+/// can deal with them later, as apropriate.
+///
+/// By memorizing the original generator function and environment, it also can be reset to its
+/// initial state by calling the original generator function again to fetch a new freshly
+/// initialized function.
+///
+/// TODO: Upvalues of generators or simple functions could actuially be collected and restored
+/// too, but this uses debug functionality and may break some upvalues.
+#[derive(Debug, Clone)]
+pub(crate) struct LuaCallback {
+    environment: Option<LuaOwnedTable>,
+    context: LuaOwnedTable,
+    generator: Option<LuaOwnedFunction>,
+    function: LuaOwnedFunction,
+    initialized: bool,
 }
 
-// -------------------------------------------------------------------------------------------------
+impl LuaCallback {
+    pub fn new(lua: &Lua, function: LuaFunction) -> LuaResult<Self> {
+        // create an empty context and memorize the function without calling it
+        let context = lua.create_table()?.into_owned();
+        let environment = function.environment().map(LuaTable::into_owned);
+        let generator = None;
+        let function = function.into_owned();
+        let initialized = false;
+        Ok(Self {
+            environment,
+            context,
+            generator,
+            function,
+            initialized,
+        })
+    }
 
-/// A Lua callback with an optional context, to generate events in rhythms or patterns from
-/// a lua function or fun iterator.
-///
-/// See [`LuaFunctionCallback`] and [`LuaGeneratorCallback`] for more information.
-///
-/// Use [`LuaCallbackFactory`] to create instances of this trait.
-pub(crate) trait LuaCallback: Debug {
     /// Sets the emitter time base context for the callback.
-    fn set_context_time_base(&mut self, time_base: &BeatTimeBase) -> LuaResult<()>;
+    pub fn set_context_time_base(&mut self, time_base: &BeatTimeBase) -> LuaResult<()> {
+        let table = self.context.to_ref();
+        table.raw_set("beats_per_min", time_base.beats_per_min)?;
+        table.raw_set("beats_per_bar", time_base.beats_per_bar)?;
+        table.raw_set("sample_rate", time_base.samples_per_sec)?;
+        Ok(())
+    }
 
     /// Sets external emitter context for the callback.
-    fn set_context_external_data(&mut self, data: &[(Cow<str>, f64)]) -> LuaResult<()>;
+    pub fn set_context_external_data(&mut self, data: &[(Cow<str>, f64)]) -> LuaResult<()> {
+        let table = self.context.to_ref();
+        for (key, value) in data {
+            table.raw_set(key as &str, *value)?;
+        }
+        Ok(())
+    }
 
     /// Sets the pulse value emitter context for the callback.
-    fn set_context_pulse_value(&mut self, pulse: PulseIterItem) -> LuaResult<()>;
+    pub fn set_context_pulse_value(&mut self, pulse: PulseIterItem) -> LuaResult<()> {
+        let table = self.context.to_ref();
+        table.raw_set("pulse_value", pulse.value)?;
+        table.raw_set("pulse_time", pulse.step_time)?;
+        Ok(())
+    }
 
     /// Sets the pulse step emitter context for the callback.
-    fn set_context_pulse_step(
+    pub fn set_context_pulse_step(
         &mut self,
         pulse_step: usize,
         pulse_time_step: f64,
         pulse_pattern_length: usize,
-    ) -> LuaResult<()>;
+    ) -> LuaResult<()> {
+        let table = self.context.to_ref();
+        table.raw_set("pulse_step", pulse_step + 1)?;
+        table.raw_set("pulse_time_step", pulse_time_step)?;
+        table.raw_set("pattern_length", pulse_pattern_length)?;
+        table.raw_set("pattern_pulse_step", pulse_step % pulse_pattern_length + 1)?;
+        Ok(())
+    }
 
     /// Sets the step emitter context for the callback.
-    fn set_context_step(&mut self, step: usize) -> LuaResult<()>;
+    pub fn set_context_step(&mut self, step: usize) -> LuaResult<()> {
+        let table = self.context.to_ref();
+        table.raw_set("step", step + 1)?;
+        Ok(())
+    }
 
     /// Sets the emitter context for the callback. Only used for function callbacks.
-    fn set_pattern_context(
+    pub fn set_pattern_context(
         &mut self,
         time_base: &BeatTimeBase,
         pulse_step: usize,
@@ -124,7 +155,7 @@ pub(crate) trait LuaCallback: Debug {
     }
 
     /// Sets the emitter context for the callback. Only used for function callbacks.
-    fn set_emitter_context(
+    pub fn set_emitter_context(
         &mut self,
         time_base: &BeatTimeBase,
         pulse: PulseIterItem,
@@ -140,15 +171,44 @@ pub(crate) trait LuaCallback: Debug {
     }
 
     /// Name of the inner function for errors. Usually will be an annonymous function.
-    fn name(&self) -> String;
+    pub fn name(&self) -> String {
+        self.function
+            .to_ref()
+            .info()
+            .name
+            .unwrap_or("annonymous function".to_string())
+    }
 
     /// Invoke the Lua function callback or generator.
     /// Returns `None` for iterators, when the iteration finished.
-    fn call(&mut self) -> LuaResult<Option<LuaValue>>;
+    pub fn call(&mut self) -> LuaResult<Option<LuaValue>> {
+        if !self.initialized {
+            self.initialized = true;
+            let function = self.function.clone();
+            let result = function.call::<_, LuaValue>(self.context.to_ref())?;
+            if let Some(inner_function) = result.as_function() {
+                // function returned a function -> is an iterator. use inner function instead.
+                let function_environment = self
+                    .function
+                    .to_ref()
+                    .environment()
+                    .map(LuaTable::into_owned);
+                let function_generator = Some(self.function.clone());
+                self.environment = function_environment;
+                self.generator = function_generator;
+                self.function = inner_function.clone().into_owned();
+            } else {
+                // function returned not a function. use this function directly.
+                self.environment = None;
+                self.generator = None;
+            }
+        }
+        Ok(Some(self.function.call(self.context.to_ref())?))
+    }
 
     /// Report a Lua callback errors. The error will be logged and usually cleared after
     /// the next callback call.
-    fn handle_error(&self, err: &LuaError) {
+    pub fn handle_error(&self, err: &LuaError) {
         log::warn!(
             "Lua callback '{}' failed to evaluate:\n{}",
             self.name(),
@@ -161,11 +221,32 @@ pub(crate) trait LuaCallback: Debug {
     }
 
     /// Reset the callback function or iterator to its initial state.
-    fn reset(&mut self) -> LuaResult<()>;
-
-    /// Create a new cloned instance of this callback. This actualy is a clone(), wrapped into
-    /// a `Box<dyn LuaCallback>`, but called 'duplicate' to avoid conflicts with Clone trait impls.
-    fn duplicate(&self) -> Box<dyn LuaCallback>;
+    pub fn reset(&mut self) -> LuaResult<()> {
+        // resetting only is necessary when we got initialized
+        if self.initialized {
+            if let Some(function_generator) = &self.generator {
+                // restore generator environment
+                if let Some(env) = &self.environment {
+                    function_generator.to_ref().set_environment(env.to_ref())?;
+                }
+                // then fetch a new fresh function from the generator
+                let value = function_generator
+                    .to_ref()
+                    .call::<_, LuaValue>(self.context.to_ref())?;
+                if let Some(function) = value.as_function() {
+                    self.function = function.clone().into_owned();
+                } else {
+                    return Err(LuaError::runtime(format!(
+                        "Failed to reset custom generator function '{}' \
+                         Expected a function as return value, got a '{}'",
+                        self.name(),
+                        value.type_name()
+                    )));
+                }
+            }
+        }
+        Ok(())
+    }
 }
 
 // --------------------------------------------------------------------------------------------------
@@ -197,13 +278,12 @@ mod test {
     }
 
     #[test]
-    fn function() -> LuaResult<()> {
+    fn callbacks() -> LuaResult<()> {
         let (lua, _) = new_test_engine(120.0, 4, 44100)?;
 
         let rhythm = lua
             .load(
                 r#"
-                require "fun"()
                 return rhythm {
                     unit = "seconds",
                     pattern = function(context) 
@@ -218,60 +298,6 @@ mod test {
                         return note
                       end
                     end
-                }
-            "#,
-            )
-            .eval::<LuaValue>()?;
-
-        let mut rhythm = rhythm
-            .as_userdata()
-            .unwrap()
-            .borrow_mut::<SecondTimeRhythm>()?;
-        let rhythm = rhythm.borrow_mut();
-        for _ in 0..2 {
-            let events = rhythm.clone().take(4).collect::<Vec<_>>();
-            rhythm.reset();
-            assert_eq!(
-                events,
-                vec![
-                    RhythmIterItem {
-                        event: Some(Event::NoteEvents(vec![Some((Note::C4).into())])),
-                        time: 0,
-                        duration: 44100
-                    },
-                    RhythmIterItem {
-                        time: 44100,
-                        event: None,
-                        duration: 44100
-                    },
-                    RhythmIterItem {
-                        time: 88200,
-                        event: Some(Event::NoteEvents(vec![Some((Note::Ds4).into())])),
-                        duration: 44100
-                    },
-                    RhythmIterItem {
-                        time: 132300,
-                        event: Some(Event::NoteEvents(vec![Some((Note::G4).into())])),
-                        duration: 44100
-                    }
-                ]
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn generator() -> LuaResult<()> {
-        let (lua, _) = new_test_engine(120.0, 4, 44100)?;
-
-        let rhythm = lua
-            .load(
-                r#"
-                require "fun"()
-                return rhythm {
-                    unit = "seconds",
-                    pattern = iter{1,0,1}:map(function(x) return x end):cycle():take(8),
-                    emit = iter{"c4", "d#4", "g4"}:cycle()
                 }
             "#,
             )
