+++ types/nerdo/library/rhythm.lua
@@ -148,10 +148,6 @@ error("Do not try to execute this file. It's just a type definition file.")
 ---  end
 ---end
 ---
------ 'fun' iterator as pulse generator
----local fun = require "fun"
----pattern = fun.rands(5, 10):map(function(x) return x / 10.0; end):take(12):cycle()
----
 ---```
 ---@field pattern Pulse[]|(fun(context: PatternContext):Pulse)|(fun(context: PatternContext):fun(context: PatternContext):Pulse)?
 ---
@@ -209,13 +205,6 @@ error("Do not try to execute this file. It's just a type definition file.")
 ---  end
 ---end
 ---
------- a fun iterator
----local fun = require "fun"
----local cmin = scale("c5", "minor")
----local degree = function(x) return cmin:degree(x) end 
----...
----emit = fun.iter({1, 5, 2, 7, 3, 5, 3}):map(degree):cycle()
----
 ----- a note pattern
 ---local pattern = require "pattern"
 ---local tritone = scale("c5", "tritone")
@@ -245,7 +234,6 @@ error("Do not try to execute this file. It's just a type definition file.")
 ---}
 ---
 -----trigger notes in an euclidean tripplet pattern
----local pattern = require "pattern"
 ---return rhythm {
 ---  unit = "1/8",
 ---  resolution = 3/2,
